requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
flake8==3.9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
msal==1.11.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pymongo==3.11.3
pyTelegramBotAPI==3.7.7
python-dotenv==0.17.0
requests==2.25.1
urllib3==1.26.4
Werkzeug==1.0.1
